feat: ratings

+++ internal/pkg/models/rating.go
@@ -30,6 +30,6 @@ import "time"
 type Rating struct {
 	ID          uint64    `json:"id"`
 	ProductID   uint64    `json:"product_id"`
-	Score       uint32    `json:"score"`
+	Score       uint64    `json:"score"`
 	UpdatedTime time.Time `json:"updated_time"`
 }
